+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationAnnotatedTypeFactory.java
@@ -747,7 +747,7 @@ public abstract class InitializationAnnotatedTypeFactory<
                 // The receiver is not initialized for this frame and the type being computed is not
                 // a LHS.
                 // Replace all annotations with the top annotation for that hierarchy.
-                type.clearPrimaryAnnotations();
+                type.clearAnnotations();
                 type.addAnnotations(qualHierarchy.getTopAnnotations());
             }