+++ framework/src/main/java/org/checkerframework/framework/type/AsSuperVisitor.java
@@ -626,7 +626,7 @@ public class AsSuperVisitor extends AbstractAtmComboVisitor<AnnotatedTypeMirror,
         // Clear the superType annotations and copy over the primary annotations before computing
         // bounds, so that the superType annotations don't override the type annotations on the
         // bounds.
-        superType.clearPrimaryAnnotations();
+        superType.clearAnnotations();
         copyPrimaryAnnos(type, superType);
 
         AnnotatedTypeMirror upperBound = visit(type.getUpperBound(), superType.getUpperBound(), p);